NetworkSystem WIP - connection and disconnection for LAN and WAN should be working

+++ src/Core/Memory/FreeListAllocator.h
@@ -20,7 +20,7 @@ namespace NK
 		[[nodiscard]] std::uint32_t Allocate();
 		void Free(std::uint32_t _val);
 		
-		static constexpr std::size_t INVALID_INDEX{ UINT32_MAX };
+		static constexpr std::uint32_t INVALID_INDEX{ UINT32_MAX };
 		
 
 	private:
